split out postgres test files
